+++ src/components/element-hider/index.ts
@@ -1,0 +1,2 @@
+export * from './emote-holder-hider';
+export * from './top-gifters-hider';