+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './chat-watcher/chat-watcher';
+export * from './element-hider';
 export * from './settings/';