+++ src/components/settings/index.tsx
@@ -1,5 +1,4 @@
-import React from 'react';
-import { createRoot } from 'react-dom/client';
+import { injectReactElement } from '@/utils/injectReactElement';
 import { MAIN_NAVBAR_TOKEN, SETTINGS_MODAL_ID } from './settings-constants';
 import SettingsModal from './settings-modal';
 
@@ -7,36 +6,17 @@ const settingsModalElement = document.createElement('div');
 settingsModalElement.id = SETTINGS_MODAL_ID;
 settingsModalElement.classList.add('h-7');
 
-const renderSettingsModal = () => {
-  createRoot(settingsModalElement).render(
-    <React.StrictMode>
-      <SettingsModal />
-    </React.StrictMode>
-  );
-};
-
-const waitForContainer = () => {
-  return new Promise((resolve) => {
-    const container = document.body.querySelector(`.${MAIN_NAVBAR_TOKEN}`);
-    if (container) {
-      resolve(container);
-    } else {
-      const observer = new MutationObserver(() => {
-        const observerContainer = document.body.querySelector(`.${MAIN_NAVBAR_TOKEN}`);
-        if (observerContainer) {
-          observer.disconnect();
-          resolve(observerContainer);
-        }
-      });
-      observer.observe(document.body, { childList: true, subtree: true });
-    }
-  });
-};
-
-waitForContainer().then((container) => {
-  const c = container as Element;
-  c.insertBefore(settingsModalElement, c.children[c.childElementCount - 1]);
-  renderSettingsModal();
+injectReactElement({
+  rootContainer: settingsModalElement,
+  reactCompontent: <SettingsModal />,
+  mode: 'watch',
+  watchSettings: {
+    getContainer: () => document.querySelector(`.${MAIN_NAVBAR_TOKEN}`),
+    getWatchedElement: () => document.getElementById(SETTINGS_MODAL_ID),
+    elementInsertFunction: (container) => {
+      container.insertBefore(settingsModalElement, container.children[container.childElementCount - 1]);
+    },
+  },
 });
 
 export default {};